requirements.txt
attrs==24.2.0
boxsdk==3.12.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
numpy==2.0.1
pandas==2.2.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
